refactor: rename discord utils to send dm utils
chore: remove debug
